Move simple into subpackage
